警告解消

+++ Engine/3d/Model.cpp
@@ -278,7 +278,7 @@ void Model::Initialize(const std::string& filename) {
 
 		//頂点バッファビュー設定
 		vbView_.BufferLocation = vertBuff_->GetGPUVirtualAddress();
-		vbView_.SizeInBytes = sizeof(VertexData) * modelData_.vertices.size();
+		vbView_.SizeInBytes = UINT(sizeof(VertexData) * modelData_.vertices.size());
 		vbView_.StrideInBytes = sizeof(VertexData);
 
 		//マッピングしてデータ転送
